 Updated requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.5.0
twilio==8.3.0
apscheduler==3.10.1
python-dotenv==1.0.